backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
pydantic==2.11.7
python-multipart==0.0.9
email-validator==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

